Finished Refactor v1
