+++ src/components/SavedBeats.js
@@ -4,48 +4,15 @@ import EdmPads from "./EdmPads";
 import Tracks from "./Tracks";
 import Visuals from "./Visuals";
 import WubPads from "./WubPads";
+import TrapPads from "./TrapPads";
 
 function SavedBeats({ sounds }) {
 
   const [playingSongs, setPlayingSongs] = useState([])
   const [globalMute, setGlobalMute] = useState(false)
 
-  const wubstepSounds = sounds.filter(sound => sound.genre === "Wubstep")
-  const trapSounds = sounds.filter(sound => sound.genre === "Trap")
-
-  const wubstepElements = wubstepSounds.map(sound => {
-    return (<BeatButton key={sound.id} sound={sound} />)
-  })
-  const trapElements = trapSounds.map(sound => {
-    return (<BeatButton key={sound.id} sound={sound} />)
-  })
-
 
   //if they are not muted, then make a player for each of them which starts upon them all being loaded
-
-  function playWubstepSong() {
-    console.log("play song")
-    fetch(`http://localhost:8003/sounds`)
-      .then(r => r.json())
-      .then(beats => {
-        const wubBeats = beats.filter(beat => (beat.genre === "Wubstep"))
-        const wubSong = wubBeats.filter(beat => (beat.isMuted === false))
-        setPlayingSongs(wubSong)
-        setGlobalMute(false)
-      })
-  }
-
-  function playTrapSong() {
-    console.log("play song")
-    fetch(`http://localhost:8003/sounds`)
-      .then(r => r.json())
-      .then(beats => {
-        const trapBeats = beats.filter(beat => (beat.genre === "Trap"))
-        const trapSong = trapBeats.filter(beat => (beat.isMuted === false))
-        setPlayingSongs(trapSong)
-        setGlobalMute(false)
-      })
-  }
 
   const muteSwitch = () => {
     setGlobalMute(true)
@@ -71,12 +38,11 @@ function SavedBeats({ sounds }) {
             setGlobalMute={setGlobalMute}
             setPlayingSongs={setPlayingSongs}
           />
-          <h1>Trap Sounds:</h1>
-          <div className="button-container">
-            {trapElements}
-            <br></br>
-            <button className="play-btn" onClick={playTrapSong}>Play Trap Song</button>
-          </div>
+          <TrapPads
+            sounds={sounds}
+            setGlobalMute={setGlobalMute}
+            setPlayingSongs={setPlayingSongs}
+          />
           {audioSources}
           <br></br>
           <div className="container">
